Upd_030521 * Class Asset updated - implementations for a few methods. * Class Buffer92 updated - added 'ExecuteRenderCommand'. * Class EntityType updated - implementations for 'RegisterScript' & 'RegisterProperty'. * Class Folder updated - implementations for 'DestroyAllChildren' & 'UnloadBlocks', 'GetBlockId', 'SetBlockId'. * Class FragmentAsset updated - implemented a few methods. * Class GfxInternal_Dx9 updated - implementations for 'SetupRenderer', 'ResetStream', 'ResetTextures', 'RenderTexturedQuad2D_2', 'RenderTexturedQuad2D_1', 'RenderFullscreenTexture'. * Class Node updated - implementation added for 'TriggerGlobalScript'. * Class SceneSaveLoad updated - implementation for unnamed method. * Class TextureAsset updated - added a few methods implementations.

+++ TOD1/Camera.cpp
@@ -166,8 +166,8 @@ void Camera::CameraMatrix::SetupMatrix(const Vector4f& pos, const Orientation& o
 
 	float orientdist = sqrtf(orient.w * orient.w + orient.x * orient.x + orient.y * orient.y + orient.z * orient.z);
 	float fovsinneg = -sinf(fov * 0.0087266462f);
-	float fovcosneg = -cos(fov * 0.0087266462f);
-	float fovcos = cos(fov * 0.0087266462f);
+	float fovcosneg = -cosf(fov * 0.0087266462f);
+	float fovcos = cosf(fov * 0.0087266462f);
 	Vector4f invorient = {
 		1.f / orientdist * orient.w,
 		1.f / orientdist * orient.x,
